+++ mobile/android/android-components/components/feature/downloads/src/test/java/mozilla/components/feature/downloads/fake/FakeDateTimeProvider.kt
@@ -7,22 +7,8 @@ package mozilla.components.feature.downloads.fake
 import mozilla.components.feature.downloads.DateTimeProvider
 
 class FakeDateTimeProvider(
-    private val startTime: Long,
-    private val currentTime: Long,
+    private val currentTime: Long = 0,
 ) : DateTimeProvider {
 
-    private var startTimeCalled = false
-
-    override fun currentTimeMillis(): Long {
-        return if (startTimeCalled) {
-            currentTime
-        } else {
-            setStartTimeCalled()
-            startTime
-        }
-    }
-
-    private fun setStartTimeCalled() {
-        startTimeCalled = true
-    }
+    override fun currentTimeMillis(): Long = currentTime
 }