+++ mobile/android/focus-android/app/src/main/java/org/mozilla/focus/downloads/DownloadService.kt
@@ -7,7 +7,7 @@ package org.mozilla.focus.downloads
 import mozilla.components.browser.state.store.BrowserStore
 import mozilla.components.concept.fetch.Client
 import mozilla.components.feature.downloads.AbstractFetchDownloadService
-import mozilla.components.feature.downloads.DateTimeProvider
+import mozilla.components.feature.downloads.DownloadEstimator
 import mozilla.components.feature.downloads.FileSizeFormatter
 import mozilla.components.support.base.android.NotificationsDelegate
 import org.mozilla.focus.ext.components
@@ -17,5 +17,5 @@ class DownloadService : AbstractFetchDownloadService() {
     override val store: BrowserStore by lazy { components.store }
     override val notificationsDelegate: NotificationsDelegate by lazy { components.notificationsDelegate }
     override val fileSizeFormatter: FileSizeFormatter by lazy { components.fileSizeFormatter }
-    override val dateTimeProvider: DateTimeProvider by lazy { components.dateTimeProvider }
+    override val downloadEstimator: DownloadEstimator by lazy { components.downloadEstimator }
 }